requirements.txt
#Flask==3.0.0
#werkzeug==3.0.0
werkzeug==2.3.0
Flask=2.3.0
requests
unidecode
Flask-Session
python-dateutil